requirements.txt
nextcord>=2.6.0
python-dotenv>=1.0.0
paramiko>=3.4.0
psutil>=5.9.0
aiohttp>=3.9.0
